moved common units to src/core folder
