+++ src/core/modules/shared/services/api-config.service.ts
