+++ src/core/modules/shared/shared.module.ts
