+++ src/modules/auth/auth.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { JwtModule } from '@nestjs/jwt';
-import { ApiConfigService } from 'src/shared/services/api-config.service';
-import { SharedModule } from 'src/shared/shared.module';
+import { ApiConfigService } from 'src/core/modules/shared/services/api-config.service';
+import { SharedModule } from 'src/core/modules/shared/shared.module';
 import { UserModule } from '../user/user.module';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
