+++ src/modules/user/user.controller.ts
@@ -1,10 +1,7 @@
 import { Body, Controller, Get, Post } from '@nestjs/common';
-import { Roles } from 'src/decorators/roles.decorator';
-import { RoleName } from 'src/types/common.types';
 import { CreateUserDto } from './dtos/create.dto';
 import { UserService } from './user.service';
 
-@Roles(RoleName.USER)
 @Controller('users')
 export class UserController {
   constructor(private readonly userService: UserService) {}
@@ -15,7 +12,6 @@ export class UserController {
   }
 
   @Get()
-  @Roles(RoleName.ADMIN)
   getAll() {
     return this.userService.getAll();
   }
